drivers: sensor: ntc_thermistor: Adds dependency to a regulator device
With a regulator compatible device the `pullup` voltage is queried with
every temperature reading from the regulator. This allows to have
non-constant `pullup` voltage and still get valid temperature readings.

Note that the implementation does expect that the regulator is already
enabled. An error code is returned if not. Use the power-domain
to handle the activation of the regulator if not always on.

Signed-off-by: Stefan Schwendeler <Stefan.Schwendeler@husqvarnagroup.com>